Feat: 상수 위치 변경 및 함수 prop 추가
