+++ src/app/mj/page.tsx
@@ -3,7 +3,8 @@
 import { useState } from 'react';
 import WriteEditModal from '@/components/WriteEditModal/WriteEditModal';
 import { WriteEditModalType } from '@/constants/writeEditModalType';
-import Button, { BUTTON_RADIUS, BUTTON_COLOR } from '@/components/buttons/Button/Button';
+import Button from '@/components/buttons/Button/Button';
+import { BUTTON_RADIUS, BUTTON_COLOR } from '@/constants/buttonTypes';
 import classNames from 'classnames/bind';
 import styles from './page.module.scss';
 
@@ -26,6 +27,10 @@ export default function Page() {
     setSelectedModal(WriteEditModalType.editMyPost);
     // setIsOpenModal(true);
   };
+
+  const handleButtonClick = () => {
+    console.log('버튼 누름');
+  };
   return (
     <div className={cn('container')}>
       <div style={{ display: 'flex', flexDirection: 'column', gap: 10, fontSize: 20 }}>
@@ -40,7 +45,11 @@ export default function Page() {
         </button>
       </div>
       {selectedModal && <WriteEditModal type={selectedModal} />}
-      <Button backgroundColor={BUTTON_COLOR.BACKGROUND_GRAY_01} border={BUTTON_RADIUS.RADIUS_8} />
+      <Button
+        backgroundColor={BUTTON_COLOR.BACKGROUND_MAIN_01}
+        radius={BUTTON_RADIUS.RADIUS_8}
+        onClick={handleButtonClick}
+      />
     </div>
   );
 }
