+++ src/components/buttons/Button/Button.module.scss
@@ -16,7 +16,7 @@ $main-01: #587df2;
 }
 
 .common-style {
-  padding: 2rem;
+  padding: 2rem 4rem;
   font-size: 1.8rem;
 }
